requirements.txt
flask
tensorflow==2.12.0
pillow
gdown
gunicorn
